fixed untracked files

+++ src/main/java/com/project/dailylog/api/service/PostService.java
@@ -1,14 +1,17 @@
 package com.project.dailylog.api.service;
 
+import com.project.dailylog.api.domain.File;
 import com.project.dailylog.api.domain.Post;
 import com.project.dailylog.api.domain.PostEditor;
 import com.project.dailylog.api.exception.InvalidPasswordException;
 import com.project.dailylog.api.exception.PostNotFoundException;
+import com.project.dailylog.api.repository.FileRepository;
 import com.project.dailylog.api.repository.PostRepository;
 import com.project.dailylog.api.request.PostCreate;
 import com.project.dailylog.api.request.PostEdit;
 import com.project.dailylog.api.request.PostSearch;
 import com.project.dailylog.api.response.PostResponse;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +20,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
@@ -27,10 +31,27 @@ public class PostService {
   private final RedisService redisService;
   @Value("${me.cache}")
   private String cacheStore;
+  private final FileService fileService;
+  private final FileRepository fileRepository;
 
-  public PostResponse save(PostCreate postCreate){
+  @Transactional
+  public PostResponse save(PostCreate postCreate) {
     Post post = new Post().toEntity(postCreate);
     postRepository.save(post);
+
+    List<File> files = new ArrayList<>();
+    for(MultipartFile file:postCreate.getFiles()){
+      String[] upload = fileService.uploadFiles(file);
+      files.add(File.builder()
+          .realFileName(upload[1])
+          .originalFileName(file.getOriginalFilename())
+          .fileSize(file.getSize())
+          .uploadPath(upload[0])
+          .build());
+    }
+    fileRepository.saveAll(files);
+    post.addFiles(files);
+
     return new PostResponse(post);
   }
 
@@ -74,6 +95,17 @@ public class PostService {
   public void delete(long postId){
     Post post = postRepository.findById(postId)
         .orElseThrow(() -> new PostNotFoundException());
+    postRepository.delete(post);
+  }
+
+  public void checkAndDelete(long postId, Integer password){
+    Post post = postRepository.findById(postId)
+        .orElseThrow(() -> new PostNotFoundException());
+
+    if(password != null && post.getPassword() != password){
+      throw new InvalidPasswordException();
+    }
+
     postRepository.delete(post);
   }
